version 1 of AIBE-XV with falcon 7b
The python file where I have done all the heavy-lifting of LLM part. This is version 1 of the experiment with only 100 MCQ of AIBE-XV has been taken to evaluate the falcon-7b-instruct model. No local finetuning and prompting techniques have been used so far. In next version, I am planning to do so.